# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "EvidenceDesk API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/sessions": {
      "post": {
        "summary": "Start Session",
        "operationId": "start_session_sessions_post",
        "parameters": [
          {
            "name": "x-service-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Service-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tickets": {
      "get": {
        "summary": "Tickets",
        "operationId": "tickets_tickets_get",
        "parameters": [
          {
            "name": "x-session-token",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Session-Token"
            }
          },
          {
            "name": "x-service-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Service-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tickets/{ticket_id}": {
      "get": {
        "summary": "Ticket",
        "operationId": "ticket_tickets__ticket_id__get",
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Ticket Id"
            }
          },
          {
            "name": "x-session-token",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Session-Token"
            }
          },
          {
            "name": "x-service-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Service-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sources/{source_id}": {
      "get": {
        "summary": "Source",
        "operationId": "source_sources__source_id__get",
        "parameters": [
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Source Id"
            }
          },
          {
            "name": "x-session-token",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Session-Token"
            }
          },
          {
            "name": "x-service-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Service-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tickets/{ticket_id}/analyze": {
      "post": {
        "summary": "Analyze Ticket",
        "operationId": "analyze_ticket_tickets__ticket_id__analyze_post",
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Ticket Id"
            }
          },
          {
            "name": "x-session-token",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Session-Token"
            }
          },
          {
            "name": "x-service-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Service-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/runs/{run_id}": {
      "get": {
        "summary": "Run",
        "operationId": "run_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          },
          {
            "name": "x-session-token",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Session-Token"
            }
          },
          {
            "name": "x-service-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Service-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/proposals/{proposal_id}/decision": {
      "post": {
        "summary": "Decide",
        "operationId": "decide_proposals__proposal_id__decision_post",
        "parameters": [
          {
            "name": "proposal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Proposal Id"
            }
          },
          {
            "name": "x-session-token",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Session-Token"
            }
          },
          {
            "name": "x-service-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Service-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/runs/{run_id}/sources/{source_id}": {
      "get": {
        "summary": "Run Source",
        "operationId": "run_source_runs__run_id__sources__source_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          },
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Source Id"
            }
          },
          {
            "name": "x-session-token",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Session-Token"
            }
          },
          {
            "name": "x-service-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Service-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AnalyzeRequest": {
        "properties": {
          "question": {
            "type": "string",
            "maxLength": 500,
            "title": "Question",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "AnalyzeRequest"
      },
      "Answer": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "answered",
              "insufficient_evidence"
            ],
            "title": "Status"
          },
          "claims": {
            "items": {
              "$ref": "#/components/schemas/Claim"
            },
            "type": "array",
            "maxItems": 4,
            "title": "Claims"
          },
          "proposed_note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1200,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Proposed Note"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status",
          "claims"
        ],
        "title": "Answer"
      },
      "AuditResponse": {
        "properties": {
          "event": {
            "type": "string",
            "title": "Event"
          },
          "proposal_id": {
            "type": "string",
            "format": "uuid",
            "title": "Proposal Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "event",
          "proposal_id",
          "created_at"
        ],
        "title": "AuditResponse"
      },
...
      "Claim": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 600,
            "minLength": 1,
            "title": "Text"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/Citation"
            },
            "type": "array",
            "maxItems": 3,
            "minItems": 1,
            "title": "Citations"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "text",
          "citations"
        ],
        "title": "Claim"
      },
      "DecisionRequest": {
        "properties": {
          "decision": {
            "type": "string",
            "enum": [
              "approve",
              "reject"
            ],
            "title": "Decision"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "decision"
        ],
        "title": "DecisionRequest"
      },
      "DecisionResponse": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "applied",
              "rejected"
            ],
            "title": "Status"
          },
          "note_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note Id"
          },
          "replayed": {
            "type": "boolean",
            "title": "Replayed"
          }
        },
        "type": "object",
        "required": [
          "status",
          "note_id",
          "replayed"
        ],
        "title": "DecisionResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "NoteResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "content",
          "created_at"
        ],
        "title": "NoteResponse"
      },
...
      "RunResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "ticket_id": {
            "type": "string",
            "format": "uuid",
            "title": "Ticket Id"
          },
          "ticket_version": {
            "type": "integer",
            "title": "Ticket Version"
          },
          "status": {
            "type": "string",
            "enum": [
              "running",
              "complete",
              "failed"
            ],
            "title": "Status"
          },
          "mode": {
            "type": "string",
            "enum": [
              "simulated",
              "live"
            ],
            "title": "Mode"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Answer"
              },
              {
                "type": "null"
              }
            ]
          },
          "proposal": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProposalResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/SourceResponse"
            },
            "type": "array",
            "title": "Evidence"
          },
          "trace": {
            "additionalProperties": true,
            "type": "object",
            "title": "Trace"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "ticket_id",
          "ticket_version",
          "status",
          "mode",
          "result",
          "proposal",
          "evidence",
          "trace",
          "error_code",
          "created_at"
        ],
        "title": "RunResponse"
      },
      "RunSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "mode": {
            "type": "string",
            "title": "Mode"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "mode",
          "created_at"
        ],
        "title": "RunSummary"
      },
      "SessionResponse": {
        "properties": {
          "token": {
            "type": "string",
            "title": "Token"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "mode": {
            "type": "string",
            "enum": [
              "simulated",
              "live"
            ],
            "title": "Mode"
          }
        },
        "type": "object",
        "required": [
          "token",
          "expires_at",
          "mode"
        ],
        "title": "SessionResponse"
      },
      "SourceResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "heading": {
            "type": "string",
            "title": "Heading"
          },
          "body": {
            "type": "string",
            "title": "Body"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "source_id": {
            "type": "string",
            "title": "Source Id"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "active"
          }
        },
        "type": "object",
        "required": [
          "id",
          "heading",
          "body",
          "title",
          "source_id",
          "version"
        ],
        "title": "SourceResponse"
      },
      "TicketResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "sample": {
            "type": "string",
            "title": "Sample"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "body": {
            "type": "string",
            "title": "Body"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "notes": {
            "items": {
              "$ref": "#/components/schemas/NoteResponse"
            },
            "type": "array",
            "title": "Notes"
          },
          "audit": {
            "items": {
              "$ref": "#/components/schemas/AuditResponse"
            },
            "type": "array",
            "title": "Audit"
          },
          "runs": {
            "items": {
              "$ref": "#/components/schemas/RunSummary"
            },
            "type": "array",
            "title": "Runs"
          }
        },
        "type": "object",
        "required": [
          "id",
          "sample",
          "title",
          "body",
          "version",
          "notes",
          "audit",
          "runs"
        ],
        "title": "TicketResponse"
      },
      "TicketSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "sample": {
            "type": "string",
            "title": "Sample"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "body": {
            "type": "string",
            "title": "Body"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          }
        },
        "type": "object",
        "required": [
          "id",
          "sample",
          "title",
          "body",
          "version"
        ],
        "title": "TicketSummary"
      },
      "TicketsResponse": {
        "properties": {
          "tickets": {
            "items": {
              "$ref": "#/components/schemas/TicketSummary"
            },
            "type": "array",
            "title": "Tickets"
          },
          "mode": {
            "type": "string",
            "enum": [
              "simulated",
              "live"
            ],
            "title": "Mode"
          }
        },
        "type": "object",
        "required": [
          "tickets",
          "mode"
        ],
        "title": "TicketsResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}